grayscale/requirements.txt
cheroot==8.5.1
CherryPy==18.6.0
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
importlib-resources==4.1.1
jaraco.classes==3.1.0
jaraco.collections==3.1.0
jaraco.functools==3.1.0
jaraco.text==3.4.0
kiwisolver==1.3.1
matplotlib==3.3.3
more-itertools==8.6.0
networkx==2.5
numpy==1.19.4
Pillow==8.1.0
portend==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tempora==4.0.1
tifffile==2020.9.3
zc.lockfile==2.0
zipp==3.4.0
